backend/requirements.txt
fastapi[all]==0.110.2         
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.32
psycopg[binary]==3.1.18
python-dotenv==1.0.1
spotipy==2.23.0
numpy==1.26.4
scikit-learn==1.3.2
redis==5.0.4
pydantic==2.8.1
websockets==12.0
httpx==0.27.0       